@@ -15,16 +15,15 @@
     "@types/swiper": "^6.0.0",
     "@types/vue-i18n": "^7.0.0",
     "@types/vue-router": "^2.0.0",
-    "@vueuse/core": "^10.1.2",
     "axios": "^1.4.0",
     "json-server": "^0.17.3",
+    "pinia": "^2.1.3",
     "prettier": "^2.8.8",
     "quasar": "^2.12.0",
     "swiper": "^9.3.1",
-    "vue": "^3.2.47",
+    "vue": "^3.3.4",
     "vue-i18n": "^9.3.0-beta.17",
-    "vue-router": "^4.1.6",
-    "vuex": "^4.1.0"
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
     "@intlify/unplugin-vue-i18n": "^0.10.0",
