@@ -1,18 +1,31 @@
 <template>
-  <div class="w-full justify-center h-[130px] px-[26px]">
+  <div
+    class="w-full justify-center h-fit px-[26px] pb-5"
+    @click="openDrawer = false"
+  >
     <div class="flex h-[69px] w-full justify-between items-center">
       <q-icon name="search" size="20px" />
       <span class="text-[#22202E] text-[24px]" @click="$router.push('/')"
         >Avion</span
       >
-      <div>
+      <q-btn
+        flat
+        round
+        dense
+        icon="menu"
+        @click.stop="openNavBar"
+        class="navBarBtn"
+      />
+      <div class="leftButtuns">
         <q-icon name="o_shopping_cart" size="20px" class="mr-[16px]" />
         <q-icon name="o_account_circle" size="20px" />
       </div>
     </div>
+
+    <nav-bar :openDrawer="openDrawer" />
     <q-separator color="#0000001A" />
     <div
-      class="w-full mt-[10px] flex space-x-0 justify-center md:space-x-[44px]"
+      class="headerBottum w-full mt-[10px] space-x-0 justify-center md:space-x-[44px]"
     >
       <q-btn flat :noCaps="true"
         ><span>{{ translate("header.categories") }}</span>
@@ -63,31 +76,47 @@
       <q-btn flat :noCaps="true" @click="$router.push('/warranty')"
         ><span>{{ translate("header.warranty") }}</span></q-btn
       >
-      <!--    <button @click="changeLocale('en')">En</button>
-      <button @click="changeLocale('ru')">Ru</button> -->
+      <button @click="useStore().setLocale('ru')">ru</button>
+      <button @click="useStore().setLocale('en')">en</button>
     </div>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from "vue";
+<script lang="ts" setup>
+import { useStore } from "../store";
 import { useTranslate } from "../utils";
+import NavBar from "./NavBar.vue";
+import { ref } from "vue";
 
-export default defineComponent({
-  name: "PageHeader",
-  setup() {
-    const translate = useTranslate().t;
-    const locale = useTranslate().locale;
-    return {
-      translate,
-      locale,
-    };
-  },
+let { t: translate } = useTranslate();
 
-  methods: {
-    changeLocale(newLocale: any) {
-      this.locale = newLocale;
-    },
-  },
-});
+let openDrawer = ref(false);
+
+const openNavBar = () => {
+  openDrawer.value = true;
+};
 </script>
+<style>
+.leftButtuns {
+  display: inline-flex;
+}
+.navBarBtn {
+  display: none !important;
+}
+.headerBottum {
+  display: flex;
+}
+
+@media screen and (max-width: 640px) {
+  .leftButtuns {
+    display: none;
+  }
+  .navBarBtn {
+    display: inline-flex !important;
+  }
+
+  .headerBottum {
+    display: none;
+  }
+}
+</style>
