@@ -8,16 +8,17 @@ module.exports = {
   theme: {
     extend: {
       screens: {
-        xl: "1350px",
+        "screen-850": "850px",
+      },
+      backgroundColor: {
+        "light-grey": "#F9F9F9",
+        "dark-primary": "#2A254B",
       },
     },
     fontFamily: {
       satoshi: "Satoshi, sans-serif",
     },
-    backgroundColor: {
-      "light-grey": "#F9F9F9",
-      "dark-primary": "#2A254B",
-    },
+
     textColor: {
       "dark-primary": "#2A254B",
       "light-grey": "#F9F9F9",
